requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
httpx>=0.25
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.0
aiosqlite>=0.19
pyjwt>=2.8
bcrypt>=4.0
